requirements.txt
alembic==1.8.1
black==22.8.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
grpcio==1.48.1
grpcio-tools==1.48.1
idna==3.3
Mako==1.2.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
protobuf==3.20.1
psycopg2-binary==2.9.3
pydantic==1.10.2
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
typing_extensions==4.3.0
